+++ apps/webapp/src/loader/loadTenant.ts
@@ -1,7 +1,7 @@
 import { cache } from 'react';
 import { getClient } from '../api/client';
 import { GET_TENANT_QUERY, TENANT_COMMON_FIELDS } from 'util/tenant';
-import { graphql } from 'api/graphql';
+import { graphql, ResultOf } from 'api/graphql';
 
 export const GET_TENANT_WITH_STATS_QUERY = graphql(
   `
@@ -17,6 +17,10 @@ export const GET_TENANT_WITH_STATS_QUERY = graphql(
   `,
   [TENANT_COMMON_FIELDS]
 );
+
+export type TenantWithStats = NonNullable<
+  ResultOf<typeof GET_TENANT_WITH_STATS_QUERY>['tenant']
+>;
 
 export class TenantNotFoundError extends Error {
   name = 'TenantNotFoundError';